#6
java/jp/co/ixuiに変更
